Update industry partner logos, move to components folder
