+++ src/components/home/diverse-teams/_styles.css
