+++ src/components/home/diverse-teams/index.tsx
