+++ src/components/home/header/cover/index.tsx
