+++ src/components/home/industry-partners/_styles.css
