+++ src/components/home/industry-partners/index.tsx
@@ -6,13 +6,12 @@ const partnerLogos = [
   'bmw.jpg',
   'daimler.jpg',
   'fb.jpg',
-  'ibm.jpg',
   'nokia.jpg',
   'sap.jpg',
   'swm.jpg',
   'adac.jpg',
+  'awtc.png',
   'infineon.jpg',
-  'man.jpg',
   'telekom.jpg',
   'tuev.jpg',
 ];
