+++ src/components/home/pitch/index.tsx
