+++ src/components/home/tool-partner/_styles.css
