+++ src/components/home/tool-partner/index.tsx
