+++ src/pages/home/index.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 
-import Header from './header/index';
-import Pitch from './pitch/index';
-import DiverseTeams from './diverse-teams';
-import ExperienceMatter from './experience-matters';
-import UserFirst from './user-first';
-import OurProgram from './our-program';
-import Adac from './adac';
-import ToolPartner from './tool-partner'
-import IndustryPartners from './industry-partners';
-import WhatsNew from './whats-new';
+import Header from '../../components/home/header/index';
+import Pitch from '../../components/home/pitch/index';
+import DiverseTeams from '../../components/home/diverse-teams';
+import ExperienceMatter from '../../components/home/experience-matters';
+import UserFirst from '../../components/home/user-first';
+import OurProgram from '../../components/home/our-program';
+import Adac from '../../components/home/adac';
+import ToolPartner from '../../components/home/tool-partner'
+import IndustryPartners from '../../components/home/industry-partners';
+import WhatsNew from '../../components/home/whats-new';
 import PickTrack from '../../components/shared/pick-track';
 import Footer from '../../components/shared/footer/index';
 import CookieWarning from '../../components/shared/cookie-warning';
