+++ src/styles.css
@@ -9,16 +9,16 @@
 @import url("./components/shared/footer/_styles.css");
 @import url("./components/shared/cookie-warning/_styles.css");
 
-@import url("./pages/home/header/_styles.css");
-@import url("./pages/home/pitch/_styles.css");
-@import url("./pages/home/diverse-teams/_styles.css");
-@import url("./pages/home/experience-matters/_styles.css");
-@import url("./pages/home/user-first/_styles.css");
-@import url('./pages/home/our-program/_styles.css');
-@import url('./pages/home/adac/_styles.css');
-@import url("./pages/home/tool-partner/_styles.css");
-@import url('./pages/home/industry-partners/_styles.css');
-@import url('./pages/home/whats-new/_styles.css');
+@import url("./components/home/header/_styles.css");
+@import url("./components/home/pitch/_styles.css");
+@import url("./components/home/diverse-teams/_styles.css");
+@import url("./components/home/experience-matters/_styles.css");
+@import url("./components/home/user-first/_styles.css");
+@import url('./components/home/our-program/_styles.css');
+@import url('./components/home/adac/_styles.css');
+@import url("./components/home/tool-partner/_styles.css");
+@import url('./components/home/industry-partners/_styles.css');
+@import url('./components/home/whats-new/_styles.css');
 
 @import url('./pages/legal-disclosure/legal-content/_styles.css');
 @import url('./pages/privacy-policy/privacy-content/_styles.css');
